Add pan and expirery to vipps just like MPO
This will make vipps more like MPO. The hope is that these extra field
are redundant, because all the needed PAN info is also contained within
the payment_token. These explicit fields should only be relevant if VTS
is unavailable and 3DS is needed. Or an issuer requires 'step-up'
- which they should not - but may do.

Pan, expire_month, expire_year could be made optional, and wrong if the
same information is contained within the payment_token. But currently
I see no argument for forcing this.

See
https://github.com/clearhaus/issues-company/issues/66#issuecomment-770691021
 for discussion.