Authorization of refunds (#210)
A new requirement from schemes means that all refunds will now be
authorized
before being prepared for clearing. This means that when a refund is
initiated
we now send relevant information from the original authorization and the
refund
request upstream to be authorized. Previously we only cleared refunds.

We are implementing this with as little impact as possible for partners
as we
have all the required information already. However since we are now
authorizing
the transaction we now get a scheme and or issuer response from the
authorization. This means that a `refund` can return new status codes.
In
addition more data will be available in the response as known from
regular
authorizations. What this means is that objects `schem_response` in the
`refund`
request response.

---------

Co-authored-by: Jesper Kjeldgaard <jk@clearhaus.com>